// SPA/caams-portal/src/app/auth/responsible-person/model/responsibleperson.ts
/**
 * Api Documentation
 * Api Documentation
 *
 * OpenAPI spec version: 1.0
 * 
 *
 * NOTE: This class is auto generated by the swagger code generator program.
 * https://github.com/swagger-api/swagger-codegen.git
 * Do not edit the class manually.
 */


export interface ResponsiblePerson {
    value(value: any): unknown;
    valid: any; 
   
    organizationOid: string;
    assetOwnerOid:string;
    loginId:string;
    assetOwnerName:string;
    assetOwnerUnit:string;
    assetOwnerDesignation:string;
    assetOwnerEmail:string;
    assetOwnerMobileNumber:string;
    assetOwnerPhotoFilePath:string;
    assetOwnerStatus:string;
}